+++ bean-searcher/src/main/java/com/ejlchina/searcher/implement/DefaultMetaResolver.java
@@ -2,6 +2,7 @@ package com.ejlchina.searcher.implement;
 
 import com.ejlchina.searcher.*;
 import com.ejlchina.searcher.bean.InheritType;
+import com.ejlchina.searcher.util.StringUtils;
 
 import java.lang.reflect.Field;
 import java.util.*;
@@ -44,6 +45,17 @@ public class DefaultMetaResolver implements MetaResolver {
         }
     }
 
+    static class FieldWrapper {
+
+        private final Field field;
+        private final DbMapping.Column column;
+
+        public FieldWrapper(Field field, DbMapping.Column column) {
+            this.field = field;
+            this.column = column;
+        }
+    }
+
     protected <T> BeanMeta<T> resolveMetadata(Class<T> beanClass) {
         DbMapping.Table table = dbMapping.table(beanClass);
         if (table == null) {
@@ -54,23 +66,38 @@ public class DefaultMetaResolver implements MetaResolver {
                 snippetResolver.resolve(table.getJoinCond()),
                 snippetResolver.resolve(table.getGroupBy()),
                 table.isDistinct());
-        // 字段解析
-        Field[] fields = getBeanFields(beanClass);
-        for (int index = 0; index < fields.length; index++) {
-            Field field = fields[index];
-            if (Modifier.isStatic(field.getModifiers())) {
-                continue;
+        // 解析实体类字段
+        FieldWrapper[] wrappers = getBeanFields(beanClass).stream()
+                .map(field -> {
+                    DbMapping.Column column = dbMapping.column(beanClass, field);
+                    return column != null ? new FieldWrapper(field, column) : null;
+                })
+                .filter(Objects::nonNull)
+                // 排序：使指定了别名的 字段排在前面
+                .sorted((w1, w2) -> {
+                    int i1 = StringUtils.isBlank(w1.column.getAlias()) ? 1 : 0;
+                    int i2 = StringUtils.isBlank(w2.column.getAlias()) ? 1 : 0;
+                    return i1 - i2;
+                })
+                .toArray(FieldWrapper[]::new);
+        // 用于校验别名是否重复
+        Set<String> checkSet = new HashSet<>();
+        for (FieldWrapper wrapper: wrappers) {
+            wrapper.field.setAccessible(true);
+            String fieldAlias = resolveAlias(wrapper.column, checkSet);
+            if (fieldAlias != null) {
+                checkSet.add(fieldAlias);
+            } else {
+                throw new SearchException("The alias [" + wrapper.column.getAlias() + "] of [" + beanClass.getName()
+                        + "." + wrapper.field.getName() + "] is already exists on other fields.");
             }
-            DbMapping.Column column = dbMapping.column(beanClass, fields[index]);
-            if (column == null) {
-                continue;
-            }
-            field.setAccessible(true);
-            SqlSnippet snippet = snippetResolver.resolve(column.getFieldSql());
-            // 注意：Oracle 数据库的别名不能以下划线开头
-            FieldMeta fieldMeta = new FieldMeta(beanMeta, field, snippet, "c_" + index,
-                    column.isConditional(), column.getOnlyOn());
-            beanMeta.addFieldMeta(field.getName(), fieldMeta);
+            SqlSnippet fieldSql = snippetResolver.resolve(wrapper.column.getFieldSql());
+            FieldMeta fieldMeta = new FieldMeta(
+                    beanMeta, wrapper.field, fieldSql, fieldAlias,
+                    wrapper.column.isConditional(),
+                    wrapper.column.getOnlyOn()
+            );
+            beanMeta.addFieldMeta(wrapper.field.getName(), fieldMeta);
         }
         if (beanMeta.getFieldCount() == 0) {
             throw new SearchException("[" + beanClass.getName() + "] is not a valid SearchBean, because there is no field mapping to database.");
@@ -78,7 +105,24 @@ public class DefaultMetaResolver implements MetaResolver {
         return beanMeta;
     }
 
-    protected Field[] getBeanFields(Class<?> beanClass) {
+    protected String resolveAlias(DbMapping.Column column, Set<String> checkSet) {
+        String alias = column.getAlias();
+        if (StringUtils.isBlank(alias)) {
+            // 注意：Oracle 数据库的别名不能以下划线开头
+            int index = checkSet.size();
+            alias = "c_" + index;
+            while (checkSet.contains(alias)) {
+                alias = "c_" + (++index);
+            }
+            return alias;
+        }
+        if (checkSet.contains(alias)) {
+            return null;
+        }
+        return alias;
+    }
+
+    protected List<Field> getBeanFields(Class<?> beanClass) {
         InheritType iType = dbMapping.inheritType(beanClass);
         List<Field> fieldList = new ArrayList<>();
         Set<String> fieldNames = new HashSet<>();
@@ -99,7 +143,7 @@ public class DefaultMetaResolver implements MetaResolver {
             }
             beanClass = beanClass.getSuperclass();
         }
-        return fieldList.toArray(new Field[0]);
+        return fieldList;
     }
 
     public SnippetResolver getSnippetResolver() {